더미 데이터 생성부 추출
